move func to scripts
