+++ templates/article/article.js
@@ -1,19 +1,16 @@
 import {
   buildBlock, decorateBlock, decorateIcons, getMetadata, toClassName,
 } from '../../scripts/lib-franklin.js';
+import {
+  preloadImage,
+} from '../../scripts/scripts.js';
 
 export default async function decorate(doc) {
   if (getMetadata('section')) {
     doc.querySelector('main').classList.add(`color-${toClassName(getMetadata('section'))}`);
   }
   // preload hero image
-  const link = document.createElement('link');
-  link.rel = 'preload';
-  link.as = 'image';
-  link.href = getMetadata('og:image');
-  link.onload = resolve;
-  link.onerror = reject;
-  document.head.append(link);
+  preloadImage(getMetadata('og:image'));
 
   const firstSection = doc.querySelector('main .section');
   firstSection.before(createSectionWithHeroBlock(
